requirements.txt
supabase>=2.0.0
openai>=1.0.0
pandas>=2.0.0
matplotlib>=3.0.0
seaborn>=0.12.0
tqdm>=4.65.0
aiohttp>=3.8.0
requests>=2.31.0
python-dotenv==1.0.0
pytz==2024.1 